feat: add user creation script
